# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UserController_getAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "id",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "phone",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "roleId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "get-all-users": {
                    "summary": "Get All Users",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "email": "user@example.com",
                          "phone": "+123456789",
                          "verified": true,
                          "active": true,
                          "image": "https://example.com/avatar.png",
                          "createdAt": "2025-06-26T18:17:10.000Z",
                          "deletedAt": "2025-06-26T18:17:10.000Z"
                        }
                      ],
                      "total": 1
                    }
                  },
                  "get-user-with-id": {
                    "summary": "Get User with id",
                    "value": {
                      "data": {
                        "id": 1,
                        "name": "John Doe",
                        "email": "user@example.com",
                        "phone": "+123456789",
                        "verified": true,
                        "active": true,
                        "image": "https://example.com/avatar.png",
                        "createdAt": "2025-06-26T18:17:10.000Z",
                        "deletedAt": "2025-06-26T18:17:10.000Z"
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UserController_getAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "id",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "phone",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "roleId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "get-all-users": {
                    "summary": "Get All Users",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "email": "user@example.com",
                          "phone": "+123456789",
                          "verified": true,
                          "active": true,
                          "image": "https://example.com/avatar.png",
                          "createdAt": "2025-06-26T18:17:10.000Z",
                          "deletedAt": "2025-06-26T18:17:10.000Z"
                        }
                      ],
                      "total": 1
                    }
                  },
                  "get-user-with-id": {
                    "summary": "Get User with id",
                    "value": {
                      "data": {
                        "id": 1,
                        "name": "John Doe",
                        "email": "user@example.com",
                        "phone": "+123456789",
                        "verified": true,
                        "active": true,
                        "image": "https://example.com/avatar.png",
                        "createdAt": "2025-06-26T18:17:10.000Z",
                        "deletedAt": "2025-06-26T18:17:10.000Z"
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Structure",
    "description": "https://xd.adobe.com/view/f4331c27-8826-49e3-8d40-3b6361c2a108-a160/",
    "version": "1.0",
    "contact": {
      "name": "Fahd Hakem",
      "url": "www.linkedin.com/in/fahd-hakem",
      "email": "fhakem75@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3030",
      "description": "Local"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Sandbox"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "ACCESS Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "REFRESH Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "PASSWORD_RESET Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VERIFY Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "test@test.com"
          },
          "phone": {
            "type": "string",
            "example": "+20 1092725145"
          },
          "fcm": {
            "type": "string"
          }
        }
      },
...
    }
  }
}